fix logger, use official winstom example for the ErrorsTransport
