+++ src/imports/logger.js
@@ -1,40 +1,19 @@
 // libraries
 import winston from 'winston'
-import errorSerializer from 'pino-std-serializers/lib/err'
-import { omit, isPlainObject, isError } from 'lodash'
+import { omit, isPlainObject, mapValues } from 'lodash'
 import Crypto from 'crypto'
-import { SPLAT } from 'triple-beam'
 
 // configs
 import ErrorsTransport from './logger/ErrorsTransport'
 import conf from '../server/server.config'
 
-const { format } = winston
-const { combine, printf, timestamp } = format
-const colorizer = format.colorize()
+import { levelConfigs } from './logger/options'
+import { extended } from './logger/formatter'
 
+const { combine, timestamp, errors } = winston.format
 const { env, logLevel } = conf
 
 console.log('Starting logger', { logLevel, env })
-
-const levelConfigs = {
-  levels: {
-    error: 0,
-    warn: 1,
-    info: 2,
-    http: 3,
-    debug: 4,
-    trace: 5
-  },
-  colors: {
-    error: 'red',
-    warn: 'yellow',
-    info: 'cyan',
-    debug: 'green',
-    trace: 'white',
-    http: 'bold green'
-  }
-}
 
 const transports = [
   new winston.transports.Console({
@@ -47,50 +26,10 @@ const logger = winston.createLogger({
   transports,
   level: logLevel,
   levels: levelConfigs.levels,
-  format: combine(
-    timestamp(),
-    format.errors({ stack: true }),
-    printf(({ level, timestamp, from, userId, message, ...rest }) => {
-      const logPayload = { message, context: rest[SPLAT] }
-      const stringifiedPayload = JSON.stringify(logPayload, (_, value) =>
-        isError(value) ? errorSerializer(value) : value
-      )
-      return colorizer.colorize(
-        level,
-        `${timestamp} - workerId:${global.workerId} - ${level}${
-          from ? ` (FROM ${from} ${userId || ''})` : ''
-        }: ${stringifiedPayload}`
-      )
-    })
-  )
+  format: combine(timestamp(), errors({ stack: true }), extended())
 })
 
 winston.addColors(levelConfigs.colors)
-
-Object.defineProperty(logger.constructor.prototype, 'async', {
-  get() {
-    let { _asyncProxy } = this
-
-    if (!_asyncProxy) {
-      const _transports = transports.filter(({ silent }) => true !== silent)
-
-      _asyncProxy = new Proxy(this, {
-        get: (target, method) => async (...args) => {
-          const promise = Promise.all(
-            _transports.map(transport => new Promise(resolve => transport.once('logged', resolve)))
-          )
-
-          target[method](...args)
-          return promise
-        }
-      })
-
-      this._asyncProxy = _asyncProxy
-    }
-
-    return _asyncProxy
-  }
-})
 
 /**
  * Sets log middleware
@@ -126,12 +65,12 @@ export const addRequestLogger = (req, res, next) => {
 // print memory fn
 const printMemory = () => {
   const used = process.memoryUsage()
-  let toPrint = {}
-  for (let key in used) {
-    toPrint[key] = `${Math.round((used[key] / 1024 / 1024) * 100) / 100} MB`
-  }
-  logger.debug('Memory usage:', toPrint)
+
+  logger.debug('Memory usage:', mapValues(used, value => `${Math.round(value / (1 << 20), 2)} MB`))
 }
-if (env !== 'test') setInterval(printMemory, 30000)
+
+if ('test' !== env) {
+  setInterval(printMemory, 30000)
+}
 
 export default logger
