+++ src/imports/logger/options.js
@@ -1,0 +1,20 @@
+// @flow
+
+export const levelConfigs = {
+  levels: {
+    error: 0,
+    warn: 1,
+    info: 2,
+    http: 3,
+    debug: 4,
+    trace: 5
+  },
+  colors: {
+    error: 'red',
+    warn: 'yellow',
+    info: 'cyan',
+    debug: 'green',
+    trace: 'white',
+    http: 'bold green'
+  }
+}
